Fix ioctl call to FE_READ_BER in input_docsis. Update SNMP MIB for perfs. Added OIDs for target perf objects. Fix xmlrpccmd_get_target() broken in last commit. Fix timestamp handling on big endian platform for datastore_postgres. Added extra perf objects to target_dump_payload. Improve locking of perf objects. Remove now useless performance stats on target_inject.

+++ src/input_docsis.c
@@ -1723,7 +1723,7 @@ static int input_update_ber_docsis(struct perf_item *itm, void *priv) {
 
 	struct input_adapt_docsis *adapt = priv;
 
-	uint16_t ber = 0;
+	uint32_t ber = 0;
 	if (ioctl(adapt->frontend_fd, FE_READ_BER, &ber) != 0)
 		return POM_ERR;
